+++ src/dto/get-post-detail.dto.ts
@@ -1,6 +1,7 @@
 import { EmojiType } from 'src/entity/common/Enums';
 import { GetPostEmojiTuple } from 'src/repository/post-emoji.query-repository';
-import { GetPostDetailHashTagTuple, GetPostDetailTuple } from 'src/repository/post.query-repository';
+import { GetPostHashTagTuple } from 'src/repository/post-hash-tag.query-repository';
+import { GetPostDetailTuple } from 'src/repository/post.query-repository';
 import { PostDetailDto, PostWriterDto } from 'src/service/post.service';
 
 export class GetPostDetailDto {
@@ -52,7 +53,7 @@ export class GetPostDetail {
     };
   }
 
-  static from(tuple: GetPostDetailTuple, hashTagTuple: GetPostDetailHashTagTuple[], emojiTuple: GetPostEmojiTuple[]) {
+  static from(tuple: GetPostDetailTuple, hashTagTuple: GetPostHashTagTuple[], emojiTuple: GetPostEmojiTuple[]) {
     const hashTags: GetHashTagDetailInfo[] = hashTagTuple.map(tag => new GetHashTagDetailInfo(tag.tagName, tag.color));
     const emojis: GetEmojiDetailInfo[] = emojiTuple.map(emoji => new GetEmojiDetailInfo(emoji.emojiCode, emoji.emojiCount, emoji.isClicked));
     return new GetPostDetail(
