+++ src/dto/get-post-list.dto.ts
@@ -1,6 +1,7 @@
 import { EmojiType } from 'src/entity/common/Enums';
 import { GetPostEmojiTuple } from 'src/repository/post-emoji.query-repository';
-import { GetPostListHashTagTuple, GetPostTuple } from 'src/repository/post.query-repository';
+import { GetPostHashTagTuple } from 'src/repository/post-hash-tag.query-repository';
+import { GetPostTuple } from 'src/repository/post.query-repository';
 import { PostListDto, PostWriterDto } from 'src/service/post.service';
 
 export class GetPostListDto {
@@ -51,7 +52,7 @@ export class GetPostList {
     };
   }
 
-  static from(tuple: GetPostTuple, hashTagTuple: GetPostListHashTagTuple[], emojiTuple: GetPostEmojiTuple[]) {
+  static from(tuple: GetPostTuple, hashTagTuple: GetPostHashTagTuple[], emojiTuple: GetPostEmojiTuple[]) {
     const hashTags: GetHashTagListInfo[] = hashTagTuple.map(tag => new GetHashTagListInfo(tag.tagName, tag.color));
     const emojis: GetEmojiListInfo[] = emojiTuple.map(emoji => new GetEmojiListInfo(emoji.emojiCode, emoji.emojiCount, emoji.isClicked));
     return new GetPostList(
