+++ src/dto/response/profile/get-profile-post.dto.ts
@@ -1,6 +1,7 @@
 import { EmojiType } from 'src/entity/common/Enums';
 import { GetPostEmojiTuple } from 'src/repository/post-emoji.query-repository';
-import { GetProfilePostListHashTagTuple, GetProfilePostTuple } from 'src/repository/profile.query-repository';
+import { GetPostHashTagTuple } from 'src/repository/post-hash-tag.query-repository';
+import { GetProfilePostTuple } from 'src/repository/profile.query-repository';
 import { ProfilePostListDto, ProfilePostWriterDto } from 'src/service/profile.service';
 
 export class GetProfilePostDto {
@@ -51,7 +52,7 @@ export class GetProfilePostList {
     };
   }
 
-  static from(tuple: GetProfilePostTuple, hashTagTuple: GetProfilePostListHashTagTuple[], emojiTuple: GetPostEmojiTuple[]) {
+  static from(tuple: GetProfilePostTuple, hashTagTuple: GetPostHashTagTuple[], emojiTuple: GetPostEmojiTuple[]) {
     const hashTags: GetProfileHashTagListInfo[] = hashTagTuple.map(tag => new GetProfileHashTagListInfo(tag.tagName, tag.color));
     const emojis: GetProfileEmojiListInfo[] = emojiTuple.map(emoji => new GetProfileEmojiListInfo(emoji.emojiCode, emoji.emojiCount, emoji.isClicked));
     return new GetProfilePostList(
